Add files via upload
Download all the files, then run the file named "proiect.sh" in the command line (terminal). After this, the app will prompt you to enter the name of a report. The default report you just downloaded is named "date.csv", so enter this name to ensure the app works correctly. Otherwise, it will indicate that it doesn't have a file to work with. After this step, the app will run smoothly by automatically calling the other scripts.